Decouple bridge and core primitives
