+++ primitives/bridge/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "evochain-primitives"
-description = "Primitives of Evochain runtime."
+name = "ep-bridge"
+description = "Bridge primitives of Evochain runtime."
 version = "0.1.0"
 authors = ["Parity Technologies <admin@parity.io>"]
 edition = "2021"
@@ -8,13 +8,9 @@ license = "GPL-3.0-or-later WITH Classpath-exception-2.0"
 
 [dependencies]
 
-fixed-hash = { version = "0.8.0", default-features = false }
-hash256-std-hasher = { version = "0.15.2", default-features = false }
-impl-codec = { version = "0.6", default-features = false }
-impl-serde = { version = "0.4.0", default-features = false }
-parity-util-mem = { version = "0.12.0", default-features = false, features = ["primitive-types"] }
-scale-info = { version = "2.9.0", default-features = false, features = ["derive", "serde"] }
-serde = { version = "1.0", default-features = false, features = ["alloc", "derive"] }
+# Core primitives
+
+ep-core = { path = "../core", default-features = false }
 
 # Bridge Dependencies
 
@@ -28,29 +24,20 @@ frame-support = { git = "https://github.com/paritytech/substrate", branch = "pol
 frame-system = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v1.0.0", default-features = false }
 sp-api = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v1.0.0", default-features = false }
 sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v1.0.0", default-features = false }
-sp-io = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v1.0.0", default-features = false }
 sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v1.0.0", default-features = false }
 sp-std = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v1.0.0", default-features = false }
-sp-trie = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v1.0.0", default-features = false }
 
 [features]
 default = ["std"]
 std = [
 	"bp-header-chain/std",
+	"bp-messages/std",
 	"bp-runtime/std",
-	"fixed-hash/std",
+    "ep-core/std",
 	"frame-support/std",
 	"frame-system/std",
-	"hash256-std-hasher/std",
-	"impl-codec/std",
-	"impl-serde/std",
-	"parity-util-mem/std",
-	"scale-info/std",
-	"serde/std",
 	"sp-api/std",
 	"sp-core/std",
-	"sp-io/std",
 	"sp-runtime/std",
 	"sp-std/std",
-	"sp-trie/std",
 ]
